Constraints

By default, the package provides a struct to handle constraints. For each constraint, the following information is saved:

The value of the right-hand side.

The type of the constraint (<=, = or >=).

The variables (indexes) and coefficients making the left-hand side of the constraint.

With respect to the type of constraint, this is specified as a symbol. Specifically, the valid values are :lessOrEq, :equal or :greaterOrEq.

So for instance, if our MILP problem had the following constraint:

\[x_1 + 2x_3 - 4x_7 + 3x_9 \leq 17 \]

We would build it as follows:

vars1   = [1, 3, 7, 9];
coeffs1 = [1.0, 2.0, -4.0, 3.0]; 
rhs1 = 17.0; 
myConstraint1 = OREnvironment.constructConstraint(rhs1, :lessOrEq, vars1, coeffs1);

Note

The coefficients and right-hand side values of the constraints must be specified as Float64 numbers. I used to have that as a template so other types (e.g, Int) could be considered. However, I changed my mind and mimic the behavior of Gurobi, which apparently (maybe I missed something) only works with Float64 numbers.

As a final example, imagine that in your problem you also had the constraint

\[5.3x_3 - 4x_4 \geq 5 \]

Then, you would have to build it as follows:

vars2   = [3, 4];
coeffs2 = [5.3, -4.0]; 
rhs2 = 5.0; 
myConstraint2 = OREnvironment.constructConstraint(rhs2, :greaterOrEq, vars2, coeffs2);

If for any reason this default constraint structure doesn't fulfill your needs, you can always define a new type of constraint while reusing most of the methods available for the default version. See further discussion later on this page.

Building constraints as in practice

In most cases, you don't build constraints one by one as shown above, but you read them from a file or some data structure. Let's see how this can accomplish in OREnvironment. We show two different ways. One will be more useful than the other depending on the situation.

Option 1

Imagine that we have the information about the constraints (variables, coefficients, etc) stored in a dataframe of N rows. Then we will create the constraints as follows:

constraints = Array{OREnvironment.Constraint, 1}(); # vector of constraints
for i in 1:N
  c = OREnvironment.constructConstraint(df.rhs[i], df.symbol[i], df.vars[i], df.coefs[i]);
  push!(constraints, c);
end

Option 2

On some occasions, you will have your constraints stored in a text file. Then, you can load the constraints as follows:

file = "NameOfTheFileWithTheConstraints.txt"
constraints = OREnvironment.read_constraints(file);

For an example (which I highly recommed you to read), check the file ./test/sampleConstraints.txt and the @testset "read_constraints" in file ./test/Constraints.jl

Note

Function OREnvironment.read_constraints() makes some important assumptions in the file to read:

In the file, the constraints are separated by a newline. That is, no comma or other character appears at the end of each line.

In the file, all the variables are named as "x". To distinguish between variables an index is added. So for instance, if your constraint were $2x + 3y + 5 z >= 12$, you should write in your file: 2x1 + 3x2 + 5x3 >= 12.

All the variables must have a coefficient multiplying them. That is, if your constraint is of type $x_1 + x_2 \leq 3$ you must write 1 as a coefficient. So your constraint in the file looks like: 1x1 + 1x2 <= 3.

All the terms in the left-hand side are adding. Negative symbols are for the numbers. So for instance, if your constraint were $4x_3 -5.3x_{12} + x_{152} = 3$, in your file you should write: 4x3 + -5.3x12 + 1x152 = 3.

Some technical details

The type of constraint considered by default is called OREnvironment.DefaultConstraint, which inherits from OREnvironment.Constraint. So you can define your own type of constraint and reuse most of the code (if not all) available for OREnvironment.Constraint by using inheritance.

An interesting feature of OREnvironment.DefaultConstraint is that, although it just asks for the variables and coefficients making the left-hand side of a constraint, it handles this information in a special way. It saves the information distinguishing between variables with a positive coefficient and variables with a negative one. This is an idea that I read from a paper by the people of LocalSolver, which is particularly useful to develop black-box solver methods. The relationship between variables and coefficients is saved in a Julia dictionary for fast access.

For more information you can always check the source code /src/Constraints.jl.

Methods for constraints

We finish this part of the documentation showing the methods available to deal with constraints.

OREnvironment.compute_lhs

OREnvironment.compute_lhs_after_increment

OREnvironment.constructConstraint

OREnvironment.get_coefficient

OREnvironment.get_relationship_variables_constraints

OREnvironment.get_rhs

OREnvironment.get_type

OREnvironment.is_active

OREnvironment.is_active

OREnvironment.is_active_under_current_consumption

OREnvironment.is_current_consumption_feasible

OREnvironment.is_feasible

OREnvironment.is_feasible

OREnvironment.is_increment_feasible

OREnvironment.is_increment_feasible

OREnvironment.is_variable

OREnvironment.read_constraints

OREnvironment.set_coefficient!

OREnvironment.set_rhs!

OREnvironment.set_type!

OREnvironment.constructConstraint — Method

constructConstraint(rhs, t, vars, coefs)

Constructs a DefaultConstraint with right-hand side rhs, type t and left-hand side given by the variables in vars and coefficients in coefs.

Note: Both, the rhs and coefs must be of type Float64.

Example

julia> OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);

OREnvironment.get_rhs — Method

get_rhs(c)

Returns the right-hand side value of constraint c

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);
julia> OREnvironment.get_rhs(c)
12.0

OREnvironment.set_rhs! — Method

set_rhs!(c, val)

Sets the right-hand side value of constraint c to val

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);
julia> OREnvironment.set_rhs!(c, 25.0); 
julia> OREnvironment.get_rhs(c)
25.0

OREnvironment.get_type — Method

get_type(c)

Returns the constraint type of c. Admisible values are :lessOrEq, :equal and :greaterOrEq.

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);
julia> OREnvironment.get_type(c)
:lessOrEq

OREnvironment.set_type! — Method

set_type!(c, val)

Sets the constraint type of c to val. Admisible values are :lessOrEq, :equal and :greaterOrEq.

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);
julia> OREnvironment.set_type!(c, :greaterOrEq);
julia> OREnvironment.get_type(c)
:greaterOrEq

OREnvironment.is_variable — Method

is_variable(c, var)

Returns true if variable var is in constraint c.

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);
julia> OREnvironment.is_variable(c, 33)
true
julia> OREnvironment.is_variable(c, 13)
false

OREnvironment.get_coefficient — Method

get_coefficient(c, var)

Returns the coefficient of variable var in constraint c. If the variable is not in the constraint it returns 0.0;

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);
julia> OREnvironment.get_coefficient(c, 33)
1.0
julia> OREnvironment.get_coefficient(c, 3)
5.0

OREnvironment.set_coefficient! — Method

set_coefficient!(c, var, coef)

Sets in constraint c coefficient coef for variable var.

Note: If the variable was not in the constraint, it is added to it. Note that the vector that relates variables and constraints in the Problem struct is not updated when calling this adding a variable using this method!

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 12, 33], [2.0, 5.0, 3.0, 1.0]);
julia> OREnvironment.set_coefficient!(c, 33, 123.0);
julia> OREnvironment.get_coefficient(c, 33)
123.0

OREnvironment.get_relationship_variables_constraints — Method

get_relationship_variables_constraints(vconstraints, nVariables)

Given an Array of constraints vconstraints and the number of variables in the problem, returns an Array where the i-th element is an Array with the indexes of the constraints where the i-th variable appears.

This function is employed in the Problem struct so it is possible to quickly identify what constraints in the problem are affected when a variable changes its value. This permits faster updates of the left-hand side value associated to a given solution.

Example

julia> c1 = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 4], [2.0, 5.0, 3.0]);
julia> c2 = OREnvironment.constructConstraint(5.0, :equal, [2, 3, 5], [9.0, 1.0, 7.0]);
julia> constraints = [c1, c2];
julia> nVariables = 5;
julia> OREnvironment.get_relationship_variables_constraints(constraints, nVariables)
5-element Array{Array{Int64,1},1}:
 [1]
 [2]
 [1, 2]
 [1]
 [2]
 # So variable 1 appears in constraint 1, variable 2 in constraint 2, 
 # variable 3 in constraints 1 and 2...

OREnvironment.read_constraints — Method

read_constraints(file)

Returns an Array of constraints. The constraints are read from file.

See example in the file with the tests: ./test/Constraints.jl.

OREnvironment.is_feasible — Method

is_feasible(c, lhs)

Given a constraint c, returns if the left-hand side consumption lhs is feasible or not.

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 4], [2.0, 5.0, 3.0]);
julia> OREnvironment.is_feasible(c, 6.0)
true
julia> OREnvironment.is_feasible(c, 16.0)
false

OREnvironment.is_active — Method

is_active(c, lhs)

Given a constraint c, returns if the left-hand side consumption lhs makes the constraint active or not.

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 4], [2.0, 5.0, 3.0]);
julia> OREnvironment.is_active(c, 6.0)
false
julia> OREnvironment.is_active(c, 12.0)
true

OREnvironment.compute_lhs_after_increment — Method

compute_lhs_after_increment(var, Δ, currentLHS, c)

For a change of Δ in the value of variable var, returns the new left-hand side consumption in constraint c given the current consumption currentLHS.

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 4], [2.0, 5.0, 3.0]);
julia> Δ = 3; currentLHS = 9.0;
julia> OREnvironment.compute_lhs_after_increment(1, Δ, currentLHS, c)
15.0 # = 2*Δ + currentLHS

OREnvironment.is_increment_feasible — Method

is_increment_feasible(var, Δ, currentLHS, c)

For a change of Δ in the value of variable var, returns if new left-hand side consumption in constraint c is feasible given the current consumption currentLHS.

Example

julia> c = OREnvironment.constructConstraint(12.0, :lessOrEq, [1, 3, 4], [2.0, 5.0, 3.0]);
julia> Δ = 3; currentLHS = 9.0;
julia> OREnvironment.is_increment_feasible(1, Δ, currentLHS, c)
false # new lhs would be 15 which larger than 12

Methods where type Solution is also involved

In these methods, an struct called solution is also involved. Check Solutions for a better understanding.

OREnvironment.is_increment_feasible — Method

is_increment_feasible(s, vconstraints, var, Δ, idxConstraints)

Given solution s, the constraints of the problem vconstraints, a change Δ (respect to the current solution s) in variable var and an Array with the constraint indexes where variable var appears, returns if the change is feasible or not.

Note: That the change is feasible DOES NOT mean that the solution is feasible. Given a feasible solution, a feasible change implies that the new solution is still feasible. But given an infeasible solution, a feasible change means that the constraints where var appears are not violated by the current change. However, as the original solution was infeasible, the new one is still infeasible. If the return of the function is false, then the solution after the change is infeasible regardless of the original feasibility situation.

See examples in the file with the tests: ./test/Constraints.jl.

OREnvironment.is_current_consumption_feasible — Method

is_current_consumption_feasible(s, vconstraints)

Returns true if the current consumption associated with solution s is feasible respect to constraints vconstraints.

Note: This function DOES NOT compute the current consumption of solution s. It just checks the last values saved in memory and compare them with the right-hand side values of the constraints. This is for efficiency reasons. Usually, given a solution s, when you locally modified it, it is possible to efficiently recompute the current consumption of the solution by just checking the constraints affected by the change. Therefore, to check feasibility it would be a waste of time to recompute the left-hand sides again instead of checking the values in memory.

If the lhs must be recomputed to check feasibility, use function is_feasible(s::OREnvironment.Solution, constraints::Array{<:OREnvironment.Constraint,1}) instead.

See examples in the file with the tests: ./test/Constraints.jl.

OREnvironment.compute_lhs — Method

compute_lhs(c, s)

Computes the consumption (left-hand side) of constraint c by solution s.

Note: It returns the value of the lhs, but it DOES NOT modified the consumption status of solution s.

See examples in the file with the tests: ./test/Constraints.jl.

OREnvironment.is_feasible — Method

is_feasible(s, vconstraints)

Returns true if the consumption incurred by solution s is feasible respect to constraints vconstraints.

Note: This function computes the left-hand sides of constraints using s, but it DOES NOT update the current consumption of solution s.

Tip: If the lhs does not have to be recomputed, then check function is_current_consumption_feasible for a better performance. Actually, that function should be preferred over is_feasible(s, vconstraints).

See examples in the file with the tests: ./test/Constraints.jl.

OREnvironment.is_active — Method

is_active(c, s)

Returns true if constraint c is active under the consumption incurred by solution s.

Note: This function computes the left-hand side of the constraint using s, but it DOES NOT update the current consumption of solution s.

Tip: If the lhs does not have to be recomputed, then check function is_active_under_current_consumption for a better performance. Actually, that function should be preferred over this one.

See examples in the file with the tests: ./test/Constraints.jl.

OREnvironment.is_active_under_current_consumption — Method

is_active_under_current_consumption(c, idx, s)

Returns true if the idx-th constraint (c) is active under the consumption incurred by solution s.

Note: This function DOES NOT compute the current consumption of solution s. It just checks the last value saved in memory and compare it with the right-hand side values of the constraint. That is why idx must be provided.

If the lhs must be recomputed, use function is_active(c::OREnvironment.Constraint, s::OREnvironment.Solution) instead.

See examples in the file with the tests: ./test/Constraints.jl.