Problem environment

Usually, in MILP, a problem is defined by an objective function and a collection of constraints. To handle this structure altogether, OREnvironment provides a struct which saves the following information:

The cost coefficients of the objective function. This is an array of Float64.

An array with the constraints of the problem. See Constraints.

An array where the i-th element is an array with the constraint indexes where the i-th variable appears. This array is used for efficiency reasons in operations and it is built automatically by the constructor calling method OREnvironment.get_relationship_variables_constraints.

The sense of the optimization problem. Valid symbols are :max and :min.

The bounds of the variables.

As an example of how to build a problem, let's assume we have a vector of costs and a vector with the constraints of the problem. See Constraints for examples and details about how to build the latter. Then:

cost = rand(6); # so we have 6 variables in the problem
c1 = OREnvironment.constructConstraint(15.0, :lessOrEq, [1,3,4,6], [2.3, 3.2, 3.1, 12.34]);
c2 = OREnvironment.constructConstraint(9.0,  :lessOrEq, [1,3,5,6], [3.3, 4.2, 4.1, 13.34]);
constraints = [c1, c2];
domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
myProblem = OREnvironment.constructProblem(cost, constraints, :max, domain) 

Methods for problem

OREnvironment.add_constraint!

OREnvironment.constructProblem

OREnvironment.gap

OREnvironment.get_constraint

OREnvironment.get_constraints

OREnvironment.get_constraints_of_variable

OREnvironment.get_cost

OREnvironment.get_lb_variable

OREnvironment.get_number_of_constraints

OREnvironment.get_number_of_variables

OREnvironment.get_obj_sense

OREnvironment.get_ub_variable

OREnvironment.is_feasible

OREnvironment.set_cost!

OREnvironment.set_lb_variable!

OREnvironment.set_obj_sense!

OREnvironment.set_ub_variable!

OREnvironment.constructProblem — Method

constructProblem(c, vconstr, objSense, vdomains)

Constructs a DefaultProblem struct by providing the vector of cost coefficients c, the vector of constraint of the problem vconstr, the objective sense, and the vector of domains of the variables.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> OREnvironment.constructProblem(cost, constraints, :max, domain)

OREnvironment.get_lb_variable — Method

get_lb_variable(p, var)

Returns the lower bound of variable var.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_lb_variable(p, 1);
0.0

OREnvironment.set_lb_variable! — Method

set_lb_variable!(p, var, val)

Set the lower bound of variable var to value val.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.set_lb_variable!(p, 1, 1.0);
julia> OREnvironment.get_lb_variable(p, 1);
1.0

OREnvironment.get_ub_variable — Method

get_ub_variable(p, var)

Returns the upper bound of variable var.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_ub_variable(p, 1);
1.0

OREnvironment.set_ub_variable! — Method

set_ub_variable!(p, var, val)

Set the upper bound of variable var to value val.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.set_ub_variable!(p, 1, 3.0);
julia> OREnvironment.get_ub_variable(p, 1);
3.0

OREnvironment.get_number_of_variables — Method

get_number_of_variables(p)

Returns the number of variables in problem p.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_number_of_variables(p);
6

OREnvironment.get_number_of_constraints — Method

get_number_of_constraints(p)

Returns the number of constraints in problem p. This does not include variables bounds.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_number_of_constraints(p);
2

OREnvironment.get_cost — Method

get_cost(p, var)

Returns the cost coefficient of variable var.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_cost(p, 1);
1.0

OREnvironment.set_cost! — Method

set_cost!(p, var, coef)

Sets coef as cost coefficient of variable var.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.set_cost!(p, 1, 1.5);
julia> OREnvironment.get_cost(p, 1);
1.5

OREnvironment.get_obj_sense — Method

get_obj_sense(p)

Returns the objetive sense of the problem.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_obj_sense(p);
:max

OREnvironment.set_obj_sense! — Method

set_obj_sense!(p, objSense)

Sets objSense as the objetive sense of the problem.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.set_obj_sense!(p, :min);
julia> OREnvironment.get_obj_sense(p);
:min

OREnvironment.get_constraint — Method

get_constraint(p, i)

Returns the i-th constraint of the problem.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_constraint(p, 1)

OREnvironment.get_constraints — Method

get_constraints(p)

Returns the constraints of the problem.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_constraints(p)

OREnvironment.get_constraints_of_variable — Method

get_constraints_of_variable(p, var)

Returns the indexes of the constraints where variable var appears in the problem.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.get_constraints_of_variable(p, 1)
[1, 2] # variable 1 appears in both constraints

OREnvironment.add_constraint! — Method

add_constraint!(p, c)

Adds constraint c to problem p. It is also possible to provide a vector of constraints instead of a single constraint.

Note: The vector indicating which variables appears in which constraints is also updated here.

Example

julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> c = OREnvironment.constructConstraint(3.9, :equal, [5, 6], [1.4, 11.0]);
julia> OREnvironment.add_constraint!(p, c);

OREnvironment.is_feasible — Method

is_feasible(p, s)

Checks if the current status of solution s is feasible for problem p, regarding constraints and variables bounds.

Note: This function DOES NOT compute the current consumption of solution s. It just checks the last values saved in memory and compare them with the right-hand side values of the constraints.

Example

julia> typeVariables = Int; numVariables = 6;  numConstraints = 2;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> cost = collect(1.0:6.0);
julia> variables1 = [1, 3, 4, 6];
julia> variables2 = [1, 3, 5, 6];
julia> coefs1 = [2.3, 3.2, 3.1, 12.34];
julia> coefs2 = coefs1 .+ 1.0;
julia> constraint1 = OREnvironment.constructConstraint(15.0, :lessOrEq, variables1, coefs1);
julia> constraint2 = OREnvironment.constructConstraint(9.0, :lessOrEq, variables2, coefs2);
julia> constraints = [constraint1, constraint2];
julia> domain = [OREnvironment.VariableDomain(0.0,1.0) for i in 1:6];
julia> p = OREnvironment.constructProblem(cost, constraints, :max, domain);
julia> OREnvironment.is_feasible(p,s)
true
julia> OREnvironment.add_solution!(s, 1, 2) # value 2 violates domain
julia> OREnvironment.is_feasible(p,s)
false

OREnvironment.gap — Method

gap(LB, UB, objSense)

Returns the gap between the lower bound LB and the upper bound UB. The computation changes depending on the objective sense of the problem.

Example

julia> LB = 7.0;
julia> UB = 9.0;
julia> OREnvironment.gap(LB, UB, :min) == (UB-LB)/LB
true
julia> OREnvironment.gap(LB, UB, :max) == (UB-LB)/UB
true