Status

Status is a struct oriented to save the current status of a given solution. That is, information related to the solution which is more efficient to save in memory than recomputing it all the time. E.g, if the solution is feasible or not.

The default struct provided can save info about:

Feasibility (boolean).

Optimality (boolean)

Value of the objective function (Float64).

Left-hand side values (a.k.a, constraint consumption). This is an array of Float64 of the same size that the number of constraints in the problem. Its purpose is to save the left-hand side values produced by a given solution.

Note that the latter is mainly oriented to save time in the following way: Given a solution and the corresponding left-hand side values in the associated status, when the solution changes, it is possible to update the left-hand side values without recomputing everything from scratch.

To create a default status, just the number of constraints has to be provided.

numberConstraints = 12;
myStatus = OREnvironment.constructStatus(numberConstraints)

Methods for status

Note

Unless your intention is to extend the default status provided in the package, forget about this section. The same methods are available for the solution struct. Check them in Solutions.

OREnvironment.constructStatus

OREnvironment.get_constraint_consumption

OREnvironment.get_objfunction

OREnvironment.is_feasible

OREnvironment.is_first_status_better

OREnvironment.is_optimal

OREnvironment.set_constraint_consumption!

OREnvironment.set_feasible!

OREnvironment.set_objfunction!

OREnvironment.set_optimal!

OREnvironment.update_status!

OREnvironment.worst_value

OREnvironment.constructStatus — Method

constructStatus(nConstraints)

Constructs a DefaultStatus. It is necessary to specify the number of constraints of the problem.

Example

julia> OREnvironment.constructStatus(5);

OREnvironment.is_feasible — Method

is_feasible(s)

Returns the feasibility value of status s.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.is_feasible(s)
false

OREnvironment.set_feasible! — Method

set_feasible!(s, val)

Sets the feasibility value of status s to val.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.set_feasible!(s, true);
julia> OREnvironment.is_feasible(s)
true

OREnvironment.is_optimal — Method

is_optimal(s)

Returns the optimality value of status s.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.is_optimal(s)
false

OREnvironment.set_optimal! — Method

set_optimal!(s, val)

Sets the optimality value of status s to val.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.set_optimal!(s, true);
julia> OREnvironment.is_optimal(s)
true

OREnvironment.get_objfunction — Method

get_objfunction(s)

Returns the objective function value of status s.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.get_objfunction(s)
0.0

OREnvironment.set_objfunction! — Method

set_objfunction!(s, val)

Sets the objective function value of status s to val.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.set_objfunction!(s, 19.8);
julia> OREnvironment.get_objfunction(s)
19.8

OREnvironment.get_constraint_consumption — Method

get_constraint_consumption(s, idx)

Returns the consumption of the idx-th constraint in status s.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.get_constraint_consumption(s, 1)
0.0

OREnvironment.set_constraint_consumption! — Method

set_constraint_consumption!(s, idx, val)

Sets the consumption of the idx-th constraint to val in status s.

Example

julia> s = OREnvironment.constructStatus(5);
julia> OREnvironment.set_constraint_consumption!(s, 2, 3.4);
julia> OREnvironment.get_constraint_consumption(s, 2)
3.4

OREnvironment.worst_value — Method

worst_value(objSense)

Returns the worst value of the objetive function according to the objective sense of the problem.

Example

julia> OREnvironment.worst_value(:max)
-Inf
julia> OREnvironment.worst_value(:min)
Inf

OREnvironment.is_first_status_better — Method

is_first_status_better(s1, s2, objSense, feasibilityRequiered)

Returns if status s1 is better than status s2. For the comparaison is required the objective function sense and if feasibility is a must for the comparaison.

Example

julia> s1 = OREnvironment.constructStatus(5);
julia> OREnvironment.set_feasible!(s1, true);
julia> OREnvironment.set_objfunction!(s1, 19.8);
julia> s2 = OREnvironment.constructStatus(5);
julia> OREnvironment.set_feasible!(s2, false);
julia> OREnvironment.set_objfunction!(s2, 100.0);
julia> OREnvironment.is_first_status_better(s1, s2, :max, true)
true
julia> OREnvironment.is_first_status_better(s1, s2, :max, false)
false
julia> OREnvironment.set_feasible!(s2, true);
julia> OREnvironment.is_first_status_better(s1, s2, :max, true)
false
julia> OREnvironment.is_first_status_better(s1, s2, :min, true)
true

Methods for DefaultStatus

If you implement a new status, you also have to implement this method.

OREnvironment.update_status! — Method

update_status!(s1, s2)

Copies info of status s2 to s1.

Example

julia> s1 = OREnvironment.constructStatus(5);
julia> s2 = OREnvironment.constructStatus(5);
julia> OREnvironment.set_feasible!(s2, true);
julia> OREnvironment.set_optimal!(s2, true);
julia> OREnvironment.set_objfunction!(s2, 100.0);
julia> OREnvironment.is_feasible(s1)
false
julia> OREnvironment.is_optimal(s1)
false
julia> OREnvironment.get_objfunction(s1)
0.0
julia> OREnvironment.update_status!(s1, s2);
julia> OREnvironment.is_feasible(s1)
true
julia> OREnvironment.is_optimal(s1)
true
julia> OREnvironment.get_objfunction(s1)
100.0