Solutions

To handle solutions, the package again provides a default struct. This saves the following information:

A given status that we want to associate with the solution. See for example the default one implemented in Status.

A solution representation. The one implemented in the package consists of an array of fixed length (i.e, it does not vary its size once created).

As an example, consider we have a problem with 12 constraints, a total of 9 variables and all of them are integer. Then:

# We first build the status for the solution
numberConstraints = 12;
myStatus = OREnvironment.constructStatus(numberConstraints);
# Then, we build the rest of the arguments required
typeOfVariables = Int;
numberOfVariables = 9; 
args = (typeOfVariables, numberOfVariables, myStatus);
# Finally, we build the solution 
mySolution = OREnvironment.constructSolution(:FixedLengthArray, args)

A couple of important notes:

If at least one of the variables is continuous, then use Float64 instead of Int.

For each solution you create, build a new status. Naturally, when dealing with multiple solutions you may want all of them to have the same type of status. However, if you simply pass as argument myStatus to all the solutions you build, Julia (probably) will create pointers. So instead of having one status per solution, you will have one status for ALL the solutions (which you don't want).

Example of what NOT to do:

numberConstraints = 12;
myStatus = OREnvironment.constructStatus(numberConstraints);
typeOfVariables = Int;
numberOfVariables = 9; 
args = (typeOfVariables, numberOfVariables, myStatus);

s1 = OREnvironment.constructSolution(:FixedLengthArray, args)
s2 = OREnvironment.constructSolution(:FixedLengthArray, args) # we are recycling the status

Example of what to do:

numberConstraints = 12;
myStatus1 = OREnvironment.constructStatus(numberConstraints);
typeOfVariables = Int;
numberOfVariables = 9; 
args = (typeOfVariables, numberOfVariables, myStatus1);

s1 = OREnvironment.constructSolution(:FixedLengthArray, args)

# See how this time we're building a new status for the new solution
myStatus2 = OREnvironment.constructStatus(numberConstraints);
args = (typeOfVariables, numberOfVariables, myStatus2);
s2 = OREnvironment.constructSolution(:FixedLengthArray, args)

About the constructor

You may wonder why in the constructor (see examples above) you have to provide :FixedLengthArray as an argument if, by default, it is the only option available.

The reason is that, when extending the package to consider other types of representations, you can build those by simply changing the arguments, not the name of the constructor.

As an example, consider that you extend the package to deal with variable length arrays. If the arguments (args) to build that solution struct are the same, the only thing of your code that should change is :FixedLengthArray by :VariableLengthArray (in case you use that identifier).

Methods for solution

OREnvironment.add_solution!

OREnvironment.add_solution_and_update_status!

OREnvironment.constructSolution

OREnvironment.create_empty_solution

OREnvironment.get_constraint_consumption

OREnvironment.get_objfunction

OREnvironment.get_solution

OREnvironment.is_feasible

OREnvironment.is_first_solution_better

OREnvironment.is_optimal

OREnvironment.remove_all_solutions!

OREnvironment.remove_all_solutions_and_update_status!

OREnvironment.remove_solution!

OREnvironment.remove_solution_and_update_status!

OREnvironment.set_constraint_consumption!

OREnvironment.set_feasible!

OREnvironment.set_objfunction!

OREnvironment.set_optimal!

OREnvironment.update_constraint_consumption!

OREnvironment.update_constraint_consumption!

OREnvironment.update_constraint_consumption_and_feasibility!

OREnvironment.update_constraint_consumption_and_feasibility!

OREnvironment.constructSolution — Method

constructSolution(solutionType, args)

Constructs a container of type solutionType for the solution. Note that the arguments of the constructor are provided in args.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)

OREnvironment.is_feasible — Method

is_feasible(s)

Returns the feasibility value of solution s.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.is_feasible(s)
false

OREnvironment.set_feasible! — Method

set_feasible!(s, val)

Sets the feasibility value of solution s to val.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.set_feasible!(s, true)
julia> OREnvironment.is_feasible(s)
true

OREnvironment.is_optimal — Method

is_optimal(s)

Returns the optimality value of solution s.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.is_optimal(s)
false

OREnvironment.set_optimal! — Method

set_optimal!(s, val)

Sets the optimality value of solution s to val.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.set_optimal!(s, true)
julia> OREnvironment.is_optimal(s)
true

OREnvironment.get_objfunction — Method

get_objfunction(s)

Returns the objective function value of solution s.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.get_objfunction(s)
0.0

OREnvironment.set_objfunction! — Method

set_objfunction!(s, val)

Sets the objective function value of solution s to val.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.set_objfunction!(s, 15.66)
julia> OREnvironment.get_objfunction(s)
15.66

OREnvironment.get_constraint_consumption — Method

get_constraint_consumption(s, idx)

Returns the consumption of the idx constraint in solution s.

Note: It returns the stored value, This method does not compute the consumption based on the current solution.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.get_constraint_consumption(s, 2)
0.0

OREnvironment.set_constraint_consumption! — Method

set_constraint_consumption!(s, idx, val)

Sets the consumption of the idx constraint in solution s to value val.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.set_constraint_consumption(s, 2, 3.4);
julia> OREnvironment.get_constraint_consumption(s, 2)
3.4

OREnvironment.is_first_solution_better — Method

is_first_solution_better(s1, s2, objSense, feasibilityRequired)

Returns if solution s1 is better than solution s2. For the comparaison is required the objective function sense and if feasibility is a must in the comparaison.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s1 = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.set_feasible!(s1, true);
julia> OREnvironment.set_objfunction!(s1, 19.8);
julia> s2 = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.set_feasible!(s2, false);
julia> OREnvironment.set_objfunction!(s2, 100.0);
julia> OREnvironment.is_first_solution_better(s1, s2, :max, true)
julia> true
julia> OREnvironment.is_first_solution_better(s1, s2, :max, false)
julia> false
julia> OREnvironment.set_feasible!(s2, true);
julia> OREnvironment.is_first_solution_better(s1, s2, :max, true)
julia> false
julia> OREnvironment.is_first_solution_better(s1, s2, :min, true)
julia> true

Methods for Solution when dealing with Constraints

OREnvironment.update_constraint_consumption! — Method

update_constraint_consumption!(s, vconstr)

Given solution s and the constraints of the problem vconstr, this method computes the left-hand side of the constraints (constraint consumption) and saves the values in memory.

See file ./test/Solution.jl for an example.

OREnvironment.update_constraint_consumption_and_feasibility! — Method

update_constraint_consumption_and_feasibility!(s, vconstr)

Given solution s and the constraints of the problem vconstr, this method computes the left-hand side of the constraints (constraint consumption), saves the values in memory and update the feasibility status according to the results obtained.

See file ./test/Solution.jl for an example.

OREnvironment.update_constraint_consumption! — Method

update_constraint_consumption!(s, vconstr, var, Δ, idxConstraints)

Updates constraint consuption of solution s for the constraints of the problem (vconstr) when variable var has changed its value by Δ. idxConstraints is the vector of constraint indexes where variable var appears.

Note that this method is more efficient than updating the constraint consumption by recomputing again all the left-hand sides of the constraints: Most of the constraints does not change when modifying a variable, and those which change, only do it in one variable! Thus, if only a few changes have occurred in the vector of solutions since the last update, this incremental method should be preferred.

See file ./test/Solution.jl for an example.

OREnvironment.update_constraint_consumption_and_feasibility! — Method

update_constraint_consumption_and_feasibility!(s, vconstr, var, Δ, idxConstraints)

Updates constraint consuption of solution s, and its feasibility status, for the constraints of the problem (vconstr) when variable var has changed its value by Δ. idxConstraints is the vector of constraint indexes where variable var appears.

Note that this method is more efficient than updating the constraint consumption by recomputing again all the left-hand sides of the constraints: Most of the constraints does not change when modifying a variable, and those which change, only do it in one variable! Thus, if only a few changes have occurred in the vector of solutions since the last update, this incremental method should be preferred.

See file ./test/Solution.jl for an example.

Methods for FixLenghtArray

These have to be reimplemented when extending the package to new solution structures.

OREnvironment.create_empty_solution — Method

create_empty_solution(problem, T)

Constructs an empty solution of type :FixedLengthArray with variables of type T with the information contained in problem (num of constraints, of variables).

OREnvironment.get_solution — Method

get_solution(s, var)

Returns the solution value of variable/index var.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.get_solution(s, 2)
0

OREnvironment.add_solution! — Method

add_solution!(s, var, val)

Sets the solution value of variable/index var to val

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.add_solution!(s, 2, 4)
julia> OREnvironment.get_solution(s, 2)
4

OREnvironment.remove_solution! — Method

remove_solution!(s, var)

Sets the solution value of variable/index var to 0.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.add_solution!(s, 2, 4)
julia> OREnvironment.get_solution(s, 2)
4
julia> OREnvironment.remove_solution!(s, 2)
julia> OREnvironment.get_solution(s, 2)
0

OREnvironment.remove_all_solutions! — Method

remove_all_solution!(s)

Sets the solution value of all the variables to 0.

Example

julia> typeVariables = Int; numVariables = 5;  numConstraints = 3;
julia> status = OREnvironment.constructStatus(numConstraints);
julia> args = (typeVariables, numVariables, status);
julia> s = OREnvironment.constructSolution(:FixedLengthArray, args)
julia> OREnvironment.add_solution!(s, 1, 12)
julia> OREnvironment.get_solution(s, 1)
12
julia> OREnvironment.add_solution!(s, 2, 4)
julia> OREnvironment.get_solution(s, 2)
4
julia> OREnvironment.remove_all_solution!(s)
julia> OREnvironment.get_solution(s, 1)
0
julia> OREnvironment.get_solution(s, 2)
0

OREnvironment.add_solution_and_update_status! — Method

add_solution_and_update_status!(s, var, val, p)

Sets the solution value of variable/index var to val and updates the status of the solution (including the consumption of the constraints in problem p).

See file ./test/Solution.jl for an example.

OREnvironment.remove_solution_and_update_status! — Method

remove_solution_and_update_status!(s, var, p)

Sets the solution value of variable/index var to 0 and updates the status of the solution (including the consumption of the constraints in problem p).

See file ./test/Solution.jl for an example.

OREnvironment.remove_all_solutions_and_update_status! — Method

remove_all_solution_and_update_status!(s, p)

Sets the solution value of all the variables to 0 and updates the status of the solution (including the consumption of the constraints in problem p).

See file ./test/Solution.jl for an example.