Movements

When developing algorithms, some operations repeatedly appear to explore the solution space. Here we provide functions for some of those so they do not have to be programmed by the user.

Methods for solution

OREnvironment.flip_value!

OREnvironment.inverse_order!

OREnvironment.mirror_value!

OREnvironment.swap_values!

OREnvironment.undo_flip!

OREnvironment.undo_inverse!

OREnvironment.undo_mirror!

OREnvironment.undo_swap!

OREnvironment.inverse_order! — Method

inverse_order!(s, var1, var2, p)

For solution s, this function reverses the order of the values between var1 and var2. For example, the inverse order of [1.2, 3.4, 6.0, 8.9] is [8.9, 6.0, 3.4, 1.2].

The status of the solution (e.g., feasibility) is updated using the information in problem p.

Example

See examples in tests.

OREnvironment.undo_inverse! — Method

undo_inverse!(s, var1, var2, p)

This function undoes the last called of function inverse_order.

Example

See examples in tests.

OREnvironment.swap_values! — Method

swap_values!(s, var1, var2, p)

For solution s, this function changes the values of variables var1 and var2, so the var1 now takes the value of var2 and var2 the value of var1.

The status of the solution (e.g., feasibility) is updated using the information in problem p.

Example

See examples in tests.

OREnvironment.undo_swap! — Method

undo_swap!(s, var1, var2, p)

This function undoes the last called of function swap_values.

Example

See examples in tests.

OREnvironment.flip_value! — Method

flip_value!(s, var, p)

For solution s, this function puts var to 0 if it has value 1 and viceversa. If the value of the variable is different than 1 or 0, then nothing happens.

The status of the solution (e.g., feasibility) is updated using the information in problem p.

Example

See examples in tests.

OREnvironment.undo_flip! — Method

undo_flip!(s, var, p)

This function undoes the last called of function flip_value.

Example

See examples in tests.

OREnvironment.mirror_value! — Method

mirror_value!(s, var, p)

If variable var in solution s has a value x[var]=z, and the variable is bounded to the interval (it may be discrete) [lb, ub], then the variable is assigned value ub+lb-z.

For example, if x[var]=4 and [lb, ub] = [1,5], then, after calling the function, x[var] = 2. Notice that if you draw the bounds in the real line: 1-2-3-4-5, value 2 is the mirror of 4. Meaning that if you fold the interval by its middle point (which in the example is 3), value 2 and 4 would touch.

If the value of the variable is not within its bounds, then nothing happens.

The status of the solution (e.g., feasibility) is updated using the information in problem p.

Example

See examples in tests.

OREnvironment.undo_mirror! — Method

undo_mirror!(s, var, p)

This function undoes the last called of function mirror_value.

Example

See examples in tests.